// typescript/src/types/openapi.d.ts
export interface OpenAPIInfo {
    readonly title: string;
    readonly version: string;
    readonly description?: string;
    readonly termsOfService?: string;
    readonly contact?: OpenAPIContact;
    readonly license?: OpenAPILicense;
    readonly summary?: string;
}
export interface OpenAPIContact {
    readonly name?: string;
    readonly url?: string;
    readonly email?: string;
}
export interface OpenAPILicense {
    readonly name: string;
    readonly identifier?: string;
    readonly url?: string;
}